Add header link to help for translations/localisation (2nd attempt)
Hi !
Here is a small update to header.php to include a link for people interested to help in translations/localisation.

The change included in this first commit is a link to the FreeCAD wiki page that has some informations about translations in FreeCAD overall (rather than a direct link to Crowdin).

Another option to discuss, proposed by @kaktusus, is to create a new page for the website, as a welcome place for new translators that could provide an in-depth description of the translation of various aspects of the FreeCAD ecosystem (GUI, docs, manual, Wiki, website, etc.), first in general and then in more detail.

What do you think ?

Note that this PR is the 2nd attempt and replaces the previous one of yesterday XD