Changing the logic of the donation code
In the old logic, onclick logic in html, the codes were nested within html, now eventlistener is used instead, this allows for easy changes without entering html. Also, the functions were edited for future reusability. Precautions were taken against multiple elements with the same id on the page. Also, a small number value in one of the previous PRs was corrected.

These changes are mostly forward-looking because it is easier to make changes to the design this way. Also, eventlisteners are necessary for doing more complex things.

Update donation.php